refactor: redesign abstract repository function structure for order domain
